Don't extend retransmitted segment
tcp_write() tries to extend the last segment in the unsent queue before
creating new segment. But the unsent queue also contains segments which
are queued for retransmission.

It is possible to queue a spurious retransmission which is the last
queued segment and therefore can be extended. If this happens, the
extended segment will be partially acknowledged and we will retransmit
duplicate data along with new data.

Although the described behavior doesn't violate RFC793, some TCP stacks
can misinterpret such a "mixed" segment. To avoid this, we don't extend
retransmitted segment when queue new data.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>